Changes done internally at Facebook
6ef99199f87aab208078ad222c78c81866c70422 Dmytro Dzhulgakov <dzhulgakov@fb.com> hacky way of fbcoding optimize-onnx
31f38d7eea59d74d225b04c638b7ce1af0626e5c Lu Fang <lufang@fb.com> Fix building issues in onnx and onnx-caffe2 to finish the import
7b84892bc9d887dacbc73f6fc3e645fe3d8a2422 Lu Fang <lufang@fb.com> Fix building issues in onnx and onnx-caffe2 to finish the import
