+++ optimize-onnx/src/export.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "ir.h"
-#include "onnx.pb.h"
+#include "onnx/onnx.pb.h"
 
 namespace onnx { namespace optimization {
